requirements.txt
fastapi
fastapi-users[sqlalchemy]
uvicorn[standard]
aiosqlite
sqladmin
python-dotenv
passlib